+++ frontend/src/app/app.module.ts
@@ -1,14 +1,23 @@
+import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { WrapperComponent } from './components/wrapper/wrapper.component';
+import { MockMainInterceptor } from './mocks/mock-main.interceptor';
 
 @NgModule({
     declarations: [AppComponent],
-    imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule],
-    providers: [],
+    imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule, WrapperComponent],
+    providers: [
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: MockMainInterceptor,
+            multi: true,
+        },
+    ],
     bootstrap: [AppComponent],
 })
 export class AppModule {}