+++ frontend/src/app/components/wrapper/wrapper.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'mados-wrapper',
+    templateUrl: './wrapper.component.html',
+    styleUrls: ['./wrapper.component.scss'],
+    standalone: true,
+})
+export class WrapperComponent {}