+++ frontend/src/app/features/main-page/components/category-type-translation/components/category-type-translation-heading/category-type-translation-heading.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+
+import { CategoryTypeTranslationHeadingComponent } from './category-type-translation-heading.component';
+
+describe('CategoryTypeTranslationHeadingComponent', () => {
+    let component: CategoryTypeTranslationHeadingComponent;
+    let fixture: ComponentFixture<CategoryTypeTranslationHeadingComponent>;
+
+    beforeEach(waitForAsync(() => {
+        TestBed.configureTestingModule({
+            declarations: [CategoryTypeTranslationHeadingComponent],
+        }).compileComponents();
+
+        fixture = TestBed.createComponent(CategoryTypeTranslationHeadingComponent);
+        component = fixture.componentInstance;
+
+        fixture.detectChanges();
+    }));
+
+    it('Should create a component', () => {
+        expect(component).withContext('Component should be created').toBeTruthy();
+    });
+});