+++ frontend/src/app/features/main-page/components/description-section/components/category/category.component.ts
@@ -1,0 +1,22 @@
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+
+import { CategoryType } from '../../../../../../enums/category-type.enum';
+
+@Component({
+    selector: 'mados-category',
+    templateUrl: './category.component.html',
+    styleUrls: ['./category.component.scss'],
+    changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class CategoryComponent {
+    @Input()
+    public type: CategoryType;
+    @Input()
+    public iconName: string;
+    @Input()
+    public description: string | null;
+    @Input()
+    public amount: number | null | undefined;
+
+    public CategoryType: typeof CategoryType = CategoryType;
+}