+++ frontend/src/app/features/main-page/components/header/header.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+
+import { HeaderComponent } from './header.component';
+
+describe('HeaderComponent', () => {
+    let component: HeaderComponent;
+    let fixture: ComponentFixture<HeaderComponent>;
+
+    beforeEach(waitForAsync(() => {
+        TestBed.configureTestingModule({
+            declarations: [HeaderComponent],
+        }).compileComponents();
+
+        fixture = TestBed.createComponent(HeaderComponent);
+        component = fixture.componentInstance;
+
+        fixture.detectChanges();
+    }));
+
+    it('Should create a component', () => {
+        expect(component).withContext('Component should be created').toBeTruthy();
+    });
+});