+++ frontend/src/app/features/main-page/services/edit-game.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+import { BehaviorSubject } from 'rxjs';
+
+@Injectable()
+export class EditGameService {
+    public editingGameName$: BehaviorSubject<string> = new BehaviorSubject<string>('');
+}