+++ frontend/src/app/interfaces/categories-response-body.interface.ts
@@ -1,0 +1,5 @@
+import { Category } from './category.interface';
+
+export interface CategoriesResponseBody {
+    categories: Category[];
+}