+++ frontend/src/app/interfaces/game-category.interface.ts
@@ -1,0 +1,6 @@
+import { CategoryType } from '../enums/category-type.enum';
+
+export interface GameCategory {
+    type: CategoryType;
+    iconName: string;
+}