+++ frontend/src/app/interfaces/game.interface.ts
@@ -1,0 +1,7 @@
+import { GameCategory } from './game-category.interface';
+
+export interface Game {
+    oneHundredPercentIconName: string;
+    name: string;
+    categories?: GameCategory[];
+}